05-monitoring/requirements.txt
pandas
numpy==1.26.4
prometheus_client
sqlalchemy 
psycopg2-binary
pytz
psutil
pyarrow
evidently